Add class for generating HOTPs
This is a very, very, very naive implementation of the algorithm
described in RFC 4226[1].

It still contains lots of debug code that I found useful in trying to
understand what was going on.

It contains a single test which uses the example given in the RFC to
assert that I'm generating this HOTP correctly.

[1]: https://www.ietf.org/rfc/rfc4226.txt