Add method to convert counter integer to 8 byte hex string
I _think_ the output of this is correct, although I can't be 100% sure.
I used trial and error to insert the counter value into the HMAC-SHA1
function to try to get the data to match that in Appendix D of RFC
4226[1]. I couldn't get it to work at all. I eventually found the
`#int_to_bytestring` method in the rotp library[2] that allowed me to
understand the required format of this string.

[1]: https://www.ietf.org/rfc/rfc4226.txt
[2]: https://github.com/mdp/rotp/blob/master/lib/rotp/otp.rb#L60