Add test to confirm my understanding of Array#pack
This test compares the result of the int_to_bytestring method from
rotp[1] to the use of `Array#pack`.

It does this by choosing 10 random numbers in the 64bit unsigned integer
space and comparing the output of the two methods. It's certainly not
foolproof but it gives me some confidence that my use of `Array#pack` is
correct.

[1]: https://github.com/mdp/rotp/blob/master/lib/rotp/otp.rb#L60