Add end-to-end tests for HOTP generation
The test data is taken from the Test Values in RFC 4226[1].

I've added a new class method that coordinates the internals of
generating an HMAC and using that to get the OTP.

I've had to change the implementation of `Hotp#hmac` to return a binary
string instead of hex representation. This is because the
`#dynamic_truncation` method expects to receive such a string. I
could've also standardised on the hex representation but binary string
seems good enough for now.

These end to end tests should give me some confidence to refactor the
internals now that I understand a bit more about the implementation.

[1]: https://www.ietf.org/rfc/rfc4226.txt