Use bit-shifting to convert byte array to integer
I tried various combinations of `Array#pack` and `String.unpack` to
convert the array of 4 bytes to a single integer but didn't get
anywhere.

I then switched to learning about bit-shifting and am now using that
instead of manually creating a string of bits and converting that to an
integer.